+++ activiti-cloud-service-common/activiti-cloud-services-test-security-keycloak/src/main/java/org/activiti/cloud/services/common/security/test/support/WithActivitiMockUser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.cloud.services.common.security.keycloak.test.support;
+package org.activiti.cloud.services.common.security.test.support;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
